requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
python-multipart==0.0.9
openpyxl==3.1.5
psycopg[binary]==3.2.9
